+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/js/userinfo/UserInfoBridgeDispatcher.kt
@@ -10,7 +10,7 @@ import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionType
  * A class to provide the interfaces for getting user info e.g. user-name, profile-photo etc.
  */
 @Suppress("TooGenericExceptionCaught", "LongMethod", "UnnecessaryAbstractClass")
-abstract class UserInfoHandler {
+abstract class UserInfoBridgeDispatcher {
 
     private lateinit var bridgeExecutor: BridgeExecutor
     private lateinit var customPermissionCache: MiniAppCustomPermissionCache